TD: adds packages, reformats, docs, pylint modifications
