+++ .gitignore
@@ -1,6 +1,9 @@
 # Custom Ignores
 cheater_scripts.txt
 
+# IDE
+.vscode/
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
