+++ run.py
@@ -1,0 +1,4 @@
+from weight_tracker import app
+
+if __name__ == "__main__":
+    app.run()
