+++ weight_tracker/templates/layout.html
