+++ weight_tracker/tests/conftest.py
@@ -8,7 +8,7 @@ root = os.path.abspath(os.path.join(os.path.dirname(__file__), ".."))
 # Add the project root directory to the system path
 sys.path.insert(0, root)
 
-from app import app as flask_app
+from run import app as flask_app
 
 
 @pytest.fixture
