+++ weight_tracker/utils/validators.py
